Add comments, use uint16 for timeout

+++ include/wdt.h
@@ -3,7 +3,7 @@
 #include <stdbool.h>
 
 void wdt_init(void);
-void wdt_set_timeout(uint8_t secs);
+void wdt_set_timeout(uint16_t secs);
 void wdt_start_reset(void);
 void wdt_stop(void);
 bool wdt_handle_timed_out(void);
